Time: 5 ms (31.78%), Space: 69.6 MB (63.07%) - LeetHub

+++ 0643-maximum-average-subarray-i/0643-maximum-average-subarray-i.java
@@ -19,6 +19,6 @@ class Solution {
             }
             j++;
         }
-        return (double)(maxa / k);
+        return (maxa / k);
     }
 }
